Extract wagmi client
