--- packages/bridge-ui/src/store/wagmi.ts
@@ -1,4 +1,0 @@
-import { writable } from 'svelte/store';
-import type { Client } from '@wagmi/core';
-
-export const wagmiClient = writable<Client>();
